refactoring employee

+++ internal/employees/domain/employee_model.go
@@ -1,9 +1,9 @@
 package domain
 
 type Employee struct {
-	ID           int64  `json:"id"`
+	ID           int    `json:"id"`
 	CardNumberId string `json:"card_number_id"`
 	FirstName    string `json:"first_name"`
 	LastName     string `json:"last_name"`
-	WarehouseId  int64  `json:"warehouse_id"`
+	WarehouseId  int    `json:"warehouse_id"`
 }
